Implement data flow, and a reaching defs routine that uses data flow
Data flow is a general technique for global (function level) analyses.
Say we want to compute certain quantities at the end of each given
block. For instance, the set of reaching definitions at the end of each
block. Given a CFG, we need to specify the following items:

1. The initial element, input to the first block. In the reaching defs
   problem, this is the set of all function arguments.
2. The transfer function for each block, which given only the input
   quantity, outputs a quantity after a local analysis of the block.
3. A merge function that, given multiple quantities, merges them into
   one.

Given these, the data flow solver converges on the desired set of
values, GIVEN that the merge function elements of the underlying poset
in one direction (or not all, but never in the reverse direction).

For instance, the meet or join operations of a (finite) poset can safely
serve as a merge function. For the reaching defs problem, where the each
element is the set of reaching variable definitions at the end of
a given block, choosing the set-union operation as the merge function
works since it is also the join operation under the subset-of relation
on the poset formed by the set of all subsets of all possible variable
defs.